requirements.txt
vanna[sqlite,gemini]>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
plotly>=5.17.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-genai>=0.3.0
pydantic>=2.0.0
